+++ crates/rslint_cli/src/files.rs
@@ -1,18 +1,19 @@
 //! The structure responsible for managing IO and the files implementation for codespan.
 
 use crate::lint_warn;
+use ignore::{WalkBuilder, WalkState};
 use rslint_core::File;
 use rslint_errors::file::{FileId, Files};
 use std::collections::HashMap;
 use std::fs::read_to_string;
 use std::ops::Range;
 use std::path::PathBuf;
-use walkdir::WalkDir;
 
-/// A list of ignored-by-default directory/file names
-const IGNORED: [&str; 1] = ["node_modules"];
 /// A list of the extension of files linted
 const LINTED_FILES: [&str; 3] = ["js", "mjs", "ts"];
+
+/// The filename of the ignore file for RSLint
+const RSLINT_IGNORE_FILE: &str = ".rslintignore";
 
 /// The structure for managing IO to and from the core runner.
 /// The walker uses multithreaded IO, spawning a thread for every file being loaded.
@@ -56,47 +57,84 @@ impl FileWalker {
 
     /// Make a new file walker from a compiled glob pattern. This also
     /// skips any unreadable files/dirs
-    pub fn from_glob(paths: Vec<PathBuf>) -> Self {
+    pub fn from_glob_parallel(paths: Vec<PathBuf>, num_threads: usize) -> Self {
         let mut base = Self::default();
-        base.load_files(paths.into_iter());
+        base.load_files_parallel(paths.into_iter(), num_threads, false, None);
         base
     }
 
-    pub fn load_files(&mut self, paths: impl Iterator<Item = PathBuf>) {
-        let jsfiles: HashMap<usize, File> = paths
-            .filter(|p| {
-                !IGNORED.contains(&p.file_name().unwrap_or_default().to_string_lossy().as_ref())
-            })
-            .flat_map(|path| {
-                WalkDir::new(path)
-                    .into_iter()
-                    .filter_entry(|p| !IGNORED.contains(&p.file_name().to_string_lossy().as_ref()))
-                    .filter_map(Result::ok)
-            })
-            .filter(|p| {
-                LINTED_FILES.contains(
-                    &p.path()
-                        .extension()
-                        .unwrap_or_default()
-                        .to_string_lossy()
-                        .as_ref(),
-                ) && !p.file_type().is_dir()
-            })
-            .filter_map(|entry| {
-                let path = entry.path();
-                let content = match read_to_string(path) {
-                    Ok(v) => v,
-                    Err(err) => {
-                        crate::lint_err!("failed to read file {}: {}", path.display(), err);
-                        return None;
+    pub fn load_files_parallel(
+        &mut self,
+        paths: impl Iterator<Item = PathBuf>,
+        num_threads: usize,
+        no_ignore: bool,
+        ignore_file: Option<PathBuf>,
+    ) {
+        let build_walker = |path: &PathBuf| {
+            let mut builder = WalkBuilder::new(path);
+            builder.standard_filters(false);
+
+            if !no_ignore {
+                match ignore_file.as_ref() {
+                    Some(file) => match builder.add_ignore(file) {
+                        Some(err) => {
+                            crate::lint_warn!("invalid gitignore file: {}", err);
+                        }
+                        None => {}
+                    },
+                    None => {
+                        builder.add_custom_ignore_filename(RSLINT_IGNORE_FILE);
                     }
-                };
-                Some((content, path.to_owned()))
-            })
-            .map(|(src, path)| File::new_concrete(src, path))
-            .map(|file| (file.id, file))
-            .collect();
-        self.files.extend(jsfiles);
+                }
+
+                builder.parents(true).hidden(true);
+            }
+
+            builder.threads(num_threads).build_parallel()
+        };
+
+        for path in paths {
+            let (tx, rx) = std::sync::mpsc::channel();
+            build_walker(&path).run(|| {
+                let tx = tx.clone();
+                Box::new(move |entry| {
+                    let path = match entry {
+                        Ok(entry) => match entry.file_type() {
+                            Some(typ) if !typ.is_dir() => entry.into_path(),
+                            _ => return WalkState::Continue,
+                        },
+                        Err(err) => {
+                            crate::lint_warn!("invalid gitignore file: {}", err);
+                            return WalkState::Continue;
+                        }
+                    };
+
+                    // check if this is a js/ts file
+                    let ext = path.extension().unwrap_or_default().to_string_lossy();
+                    if !LINTED_FILES.contains(&ext.as_ref()) {
+                        return WalkState::Continue;
+                    }
+
+                    // read the content of the file
+                    let content = match std::fs::read_to_string(&path) {
+                        Ok(c) => c,
+                        Err(err) => {
+                            crate::lint_err!("failed to read file {}: {}", path.display(), err);
+                            return WalkState::Continue;
+                        }
+                    };
+
+                    tx.send(File::new_concrete(content, path))
+                        .expect("failed to send files to receiver thread");
+                    WalkState::Continue
+                })
+            });
+
+            drop(tx);
+
+            self.files
+                .extend(rx.into_iter().map(|file| (file.id, file)));
+        }
     }
 
     pub fn line_start(&self, id: usize, line_index: usize) -> Option<usize> {