+++ crates/rslint_cli/src/infer.rs
@@ -19,7 +19,7 @@ pub fn infer(files: Vec<String>) {
         .flat_map(Result::ok)
         .collect();
 
-    let walker = FileWalker::from_glob(globs);
+    let walker = FileWalker::from_glob_parallel(globs, 1);
     let parsed = walker.files.values().map(|f| f.parse());
     let nodes: Vec<SyntaxNode> = parsed.flat_map(|n| n.descendants()).collect();
     let rules = get_inferable_rules();