+++ crates/rslint_cli/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Default support for `.rslintignore` file
+- Add `--no-ignore` and `--ignore-file` options
+
 ## [0.3.0] - 2021-09-16
 
 ### Changed